[tests][unit] Migrated assoc tests under one dir
